gstexiftag.c ascii data is written as utf8 strings.
I investigated the issue and it happens that EXIF
standard is mute on other character encodings
outside 7bit ASCII. This is largely ignored
by the industry, as in the case of gstreamer,
that uses 8 bit ASCII/latin1 encoding.

With the above, encoding the tags in UTF8 would
not result in a visible difference from what is
implemented now due to UTF8 being equivalent to
latin1 in the single byte range.
The differences would only be spotted for
characters outside the ASCII range which
previously was not supported.

Another argument to make gstreamer exif
implementation write UTF8 tags is
that it already supports parsing UTF8 encoded
tags (see gstexiftag.c@parse_exif_ascii_tag),
as verified by tests.